# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: HausHaltsMeister API
  version: 0.1.0
  description: Ledger-based finance API with budgeting, investments, and credit card flows.
servers:
  - url: /
...
security:
  - BearerAuth: []
paths:
...
  /ledgers:
    get:
      tags: [Ledgers]
      summary: List ledgers
      responses:
        '200':
          description: Ledgers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerWithRole'
    post:
      tags: [Ledgers]
      summary: Create ledger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerCreateRequest'
      responses:
        '201':
          description: Ledger
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ledger'
  /ledgers/{ledgerId}:
    get:
      tags: [Ledgers]
      summary: Get ledger
      parameters:
        - $ref: '#/components/parameters/LedgerId'
      responses:
        '200':
          description: Ledger with role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerWithRole'
    patch:
      tags: [Ledgers]
      summary: Update ledger
      parameters:
        - $ref: '#/components/parameters/LedgerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerUpdateRequest'
      responses:
        '200':
          description: Ledger
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ledger'
    delete:
      tags: [Ledgers]
      summary: Delete ledger (soft)
      parameters:
        - $ref: '#/components/parameters/LedgerId'
      responses:
        '501':
          description: Not implemented
  /ledgers/{ledgerId}/members:
    get:
      tags: [Ledgers]
      summary: List members
      parameters:
        - $ref: '#/components/parameters/LedgerId'
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerMember'
    post:
      tags: [Ledgers]
      summary: Add member
      parameters:
        - $ref: '#/components/parameters/LedgerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerMemberCreateRequest'
      responses:
        '201':
          description: Member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerMember'
  /ledgers/{ledgerId}/members/{userId}:
    patch:
      tags: [Ledgers]
      summary: Update member role
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerMemberUpdateRequest'
      responses:
        '200':
          description: Member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerMember'
    delete:
      tags: [Ledgers]
      summary: Remove member
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Removed
...
  /ledgers/{ledgerId}/reports/balances:
    get:
      tags: [Relatorios]
      summary: Account balances
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - in: query
          name: month
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Balances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceReport'
  /ledgers/{ledgerId}/reports/categories:
    get:
      tags: [Relatorios]
      summary: Category summary
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Category report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryReport'
  /ledgers/{ledgerId}/reports/cashflow:
    get:
      tags: [Relatorios]
      summary: Cashflow
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Cashflow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashflowReport'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LedgerId:
      in: path
      name: ledgerId
      required: true
      schema:
        type: string
    UserId:
      in: path
      name: userId
      required: true
      schema:
        type: string
...
  schemas:
...
    Ledger:
      type: object
      properties:
        id:
          type: string
        owner_user_id:
          type: string
        name:
          type: string
        currency_code:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    LedgerWithRole:
      type: object
      properties:
        ledger:
          $ref: '#/components/schemas/Ledger'
        role:
          type: string
    LedgerMember:
      type: object
      properties:
        ledger_id:
          type: string
        user_id:
          type: string
        role:
          type: string
        display_name:
          type: string
        email:
          type: string
        avatar_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    LedgerCreateRequest:
      type: object
      properties:
        name:
          type: string
        currency_code:
          type: string
        create_default_accounts:
          type: boolean
        include_investment:
          type: boolean
    LedgerUpdateRequest:
      type: object
      properties:
        name:
          type: string
    LedgerMemberCreateRequest:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
        role:
          type: string
    LedgerMemberUpdateRequest:
      type: object
      properties:
        role:
          type: string
...
    BalanceReport:
      type: object
      properties:
        month:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccountBalance'
    AccountBalance:
      type: object
      properties:
        account_id:
          type: string
        account_name:
          type: string
        account_type:
          type: string
        account_nature:
          type: string
        balance_cents:
          type: integer
    CategoryReport:
      type: object
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummary'
    CategorySummary:
      type: object
      properties:
        category_id:
          type: string
        name:
          type: string
        direction:
          type: string
        total_cents:
          type: integer
    CashflowReport:
      type: object
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/CashflowEntry'
    CashflowEntry:
      type: object
      properties:
        month:
          type: string
          format: date
        total_in_cents:
          type: integer
        total_out_cents:
          type: integer
        net_cents:
          type: integer