# internal/adapters/postgres/sqlc/accounts/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package accounts

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Account struct {
	ID        pgtype.UUID
	LedgerID  pgtype.UUID
	Name      string
	Type      string
	IsActive  bool
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
	Nature    string
}

type AuditLog struct {
	ID        pgtype.UUID
	LedgerID  pgtype.UUID
	UserID    pgtype.UUID
	Action    string
	EntityID  pgtype.UUID
	Ip        pgtype.Text
	UserAgent pgtype.Text
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type AuthIdentity struct {
	ID             pgtype.UUID
	UserID         pgtype.UUID
	Provider       string
	ProviderUserID string
	Email          pgtype.Text
	DisplayName    pgtype.Text
	AvatarUrl      pgtype.Text
	EmailVerified  bool
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
	LastLoginAt    pgtype.Timestamptz
}

type AuthSecret struct {
	UserID       pgtype.UUID
	PasswordHash string
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}

type AuthSession struct {
	ID                   pgtype.UUID
	UserID               pgtype.UUID
	RefreshTokenHash     string
	IsPersistent         bool
	CreatedAt            pgtype.Timestamptz
	UpdatedAt            pgtype.Timestamptz
	ExpiresAt            pgtype.Timestamptz
	RevokedAt            pgtype.Timestamptz
	UserAgent            pgtype.Text
	Ip                   pgtype.Text
	DeviceName           pgtype.Text
	RotatedFromSessionID pgtype.UUID
}

type BudgetPlan struct {
	ID        pgtype.UUID
	LedgerID  pgtype.UUID
	Name      string
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type BudgetPlanLine struct {
	ID              pgtype.UUID
	VersionID       pgtype.UUID
	CategoryID      pgtype.UUID
	Percent         pgtype.Numeric
	IncludeChildren bool
	CreatedAt       pgtype.Timestamptz
	UpdatedAt       pgtype.Timestamptz
}

type BudgetPlanVersion struct {
	ID                 pgtype.UUID
	PlanID             pgtype.UUID
	EffectiveFromMonth pgtype.Date
	CreatedByUserID    pgtype.UUID
	CreatedAt          pgtype.Timestamptz
	UpdatedAt          pgtype.Timestamptz
}

type CardNetwork struct {
	Code        string
	DisplayName string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
}

type Category struct {
	ID               pgtype.UUID
	LedgerID         pgtype.UUID
	ParentID         pgtype.UUID
	Name             string
	Direction        string
	IsBudgetBase     bool
	IsBudgetRelevant bool
	IsActive         bool
	CreatedAt        pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
}

type CreditCard struct {
	ID                 pgtype.UUID
	LedgerID           pgtype.UUID
	ParentAccountID    pgtype.UUID
	LiabilityAccountID pgtype.UUID
	Label              pgtype.Text
	Brand              string
	Last4              pgtype.Text
	Cvv                pgtype.Text
	HolderName         pgtype.Text
	Active             bool
	Color              pgtype.Text
	Style              pgtype.Text
	ClosingDay         int32
	DueDay             int32
	CreatedAt          pgtype.Timestamptz
	UpdatedAt          pgtype.Timestamptz
}

type CreditCardStatement struct {
	ID                   pgtype.UUID
	LedgerID             pgtype.UUID
	StatementMonth       pgtype.Date
	ClosingDate          pgtype.Date
	DueDate              pgtype.Date
	TotalChargesCents    int64
	TotalPaymentsCents   int64
	Status               string
	PaymentTransactionID pgtype.UUID
	CreatedAt            pgtype.Timestamptz
	UpdatedAt            pgtype.Timestamptz
	CreditCardID         pgtype.UUID
}

type Entry struct {
	ID            pgtype.UUID
	LedgerID      pgtype.UUID
	TransactionID pgtype.UUID
	AccountID     pgtype.UUID
	CategoryID    pgtype.UUID
	Kind          string
	AmountCents   int64
	Memo          pgtype.Text
	CreatedAt     pgtype.Timestamptz
	UpdatedAt     pgtype.Timestamptz
}

type IdempotencyKey struct {
	LedgerID     pgtype.UUID
	Key          string
	ResourceType string
	ResourceID   pgtype.UUID
	RequestHash  string
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
	ExpiresAt    pgtype.Timestamptz
}

type Installment struct {
	ID                  pgtype.UUID
	LedgerID            pgtype.UUID
	PlanID              pgtype.UUID
	InstallmentNo       int32
	DueMonth            pgtype.Date
	AmountCents         int64
	Status              string
	PostedTransactionID pgtype.UUID
	PaidStatementID     pgtype.UUID
	CreatedAt           pgtype.Timestamptz
	UpdatedAt           pgtype.Timestamptz
}

type InstallmentPlan struct {
	ID                     pgtype.UUID
	LedgerID               pgtype.UUID
	PurchaseOccurredAt     pgtype.Timestamptz
	Merchant               pgtype.Text
	Description            string
	CategoryID             pgtype.UUID
	TotalAmountCents       int64
	InstallmentsCount      int32
	InstallmentAmountCents int64
	FirstDueMonth          pgtype.Date
	Status                 string
	CreatedByUserID        pgtype.UUID
	CreatedAt              pgtype.Timestamptz
	UpdatedAt              pgtype.Timestamptz
	CreditCardID           pgtype.UUID
}

type Ledger struct {
	ID           pgtype.UUID
	OwnerUserID  pgtype.UUID
	Name         string
	CurrencyCode string
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}

type LedgerMember struct {
	LedgerID  pgtype.UUID
	UserID    pgtype.UUID
	Role      string
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
	RemovedAt pgtype.Timestamptz
}

type OauthState struct {
	ID           pgtype.UUID
	Provider     string
	State        string
	CodeVerifier string
	RedirectUri  pgtype.Text
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
	ExpiresAt    pgtype.Timestamptz
	UsedAt       pgtype.Timestamptz
}

type Transaction struct {
	ID              pgtype.UUID
	LedgerID        pgtype.UUID
	OccurredAt      pgtype.Timestamptz
	Description     string
	Notes           pgtype.Text
	CreatedByUserID pgtype.UUID
	ExternalSource  pgtype.Text
	ExternalID      pgtype.Text
	CreatedAt       pgtype.Timestamptz
	UpdatedAt       pgtype.Timestamptz
	CreditCardID    pgtype.UUID
}

type User struct {
	ID              pgtype.UUID
	Email           string
	DisplayName     pgtype.Text
	AvatarUrl       pgtype.Text
	EmailVerifiedAt pgtype.Timestamptz
	IsActive        bool
	CreatedAt       pgtype.Timestamptz
	UpdatedAt       pgtype.Timestamptz
}

type UserPreference struct {
	UserID           pgtype.UUID
	ThemeMode        string
	ThemePalette     string
	ThemeTone        string
	Locale           string
	CompactMode      string
	FontScale        string
	NotifyCardClose  bool
	NotifyBudgetOver bool
	NotifyPayables   bool
	CreatedAt        pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
	DefaultLedgerID  pgtype.UUID
}